added transition using framer-motion

+++ src/app/compnents/Post.js
@@ -54,11 +54,12 @@ export const Post = ({ post }) => {
     }
   }
   const deletePost = async () => {
-    if(window.confirm("Are you sure to delete this post?")){
+    if (window.confirm("Are you sure to delete this post?")) {
       deleteDoc(doc(db, "posts", post.id));
-      deleteObject(ref(storage, `posts/${post.id}/image`));
+      if (post.data().image) {
+        deleteObject(ref(storage, `posts/${post.id}/image`));
+      }
     }
-   
   };
   return (
     <div className="flex p-3 cursor-pointer border-b border-gray-200">
@@ -67,22 +68,25 @@ export const Post = ({ post }) => {
         width={100}
         height={100}
         className="h-11 w-11 rounded-full mr-4"
-        alt="post-image"
+        alt="user-image"
       />
 
-      <div className="">
+      <div className="flex-1">
+        {/* Header */}
         <div className="flex items-center justify-between">
-          <div className="flex space-x-1 items-center whitespace-nowrap">
+          {/* post user info */}
+          <div className="flex items-center space-x-1 whitespace-nowrap">
             <h4 className="font-bold text-[15px] sm:text-[16px] hover:underline">
               {post.data().name}
             </h4>
             <span className="text-sm sm:text-[15px]">
-              @{post.data().username} -
+              @{post.data().username} -{" "}
             </span>
             <span className="text-sm sm:text-[15px] hover:underline">
               <Moment fromNow>{post?.data().timestamp?.toDate()}</Moment>
             </span>
           </div>
+          {/* dot icon */}
           <EllipsisHorizontalIcon className="h-10 hoverEffect w-10 hover:bg-sky-100 hover:text-sky-500 p-2" />
         </div>
         {/* post text */}
@@ -90,13 +94,16 @@ export const Post = ({ post }) => {
           {post.data().text}
         </p>
         {/* post img */}
-        <Image
-          src={post.data().image}
-          width={500}
-          height={500}
-          className="rounded-2xl mr-2"
-          alt="post-image"
-        />
+        {post?.data().image && (
+          <Image
+            src={post.data().image}
+            width={500}
+            height={500}
+            className="rounded-2xl mr-2 h-auto w-auto"
+            alt="post-image"
+          />
+        )}
+
         {/* icons */}
         <div className="flex items-center justify-between text-gray-500 p-2">
           <ChatBubbleOvalLeftEllipsisIcon className="h-9 w-9 hoverEffect p-2 hover:text-sky-500 hover:bg-sky-100" />
